refactor: Extract types to their own file.
