+++ momento/cache_requests.go
@@ -2,27 +2,6 @@ package momento
 
 import "time"
 
-// Bytes generic interface to help users deal with passing us bytes
-type Bytes interface{ AsBytes() []byte }
-
-// RawBytes plain old []byte
-type RawBytes struct {
-	Bytes []byte
-}
-
-func (r RawBytes) AsBytes() []byte {
-	return r.Bytes
-}
-
-// StringBytes string type that will be converted to []byte
-type StringBytes struct {
-	Text string
-}
-
-func (r StringBytes) AsBytes() []byte {
-	return []byte(r.Text)
-}
-
 type CreateCacheRequest struct {
 	// string used to create a cache.
 	CacheName string
